Move maybe_convert_timedelta back out of PeriodArraY
